Export features from AttributionGraph
